refactor user_environment modification calculations
