+++ lib/spack/spack/cmd/unload.py
@@ -10,6 +10,7 @@ import spack.cmd
 import spack.cmd.common.arguments as arguments
 import spack.environment as ev
 import spack.util.environment
+import spack.user_environment as uenv
 
 description = "remove package from the user environment variables`"
 section = "user environment"
@@ -54,7 +55,8 @@ def unload(parser, args):
 
     env_mod = spack.util.environment.EnvironmentModifications()
     for spec in specs:
-        env_mod.extend(ev.environment_modifications_for_spec(spec).reversed())
+        env_mod.extend(
+            uenv.environment_modifications_for_spec(spec).reversed())
     cmds = env_mod.shell_modifications()
 
     sys.stdout.write(cmds)
