+++ lib/spack/spack/environment.py
@@ -30,7 +30,7 @@ import spack.util.spack_json as sjson
 import spack.util.spack_yaml as syaml
 import spack.config
 import spack.build_environment as build_env
-
+import spack.user_environment as uenv
 from spack.util.prefix import Prefix
 from spack.filesystem_view import YamlFilesystemView
 import spack.util.environment
@@ -1077,12 +1077,12 @@ class Environment(object):
             # No default view to add to shell
             return env_mod.shell_modifications(shell)
 
-        env_mod.extend(unconditional_environment_modifications(
+        env_mod.extend(uenv.unconditional_environment_modifications(
             self.default_view))
 
         for _, spec in self.concretized_specs():
             if spec in self.default_view and spec.package.installed:
-                env_mod.extend(environment_modifications_for_spec(
+                env_mod.extend(uenv.environment_modifications_for_spec(
                     spec, self.default_view))
 
         # deduplicate paths from specs mapped to the same location
@@ -1098,13 +1098,13 @@ class Environment(object):
             # No default view to add to shell
             return env_mod.shell_modifications(shell)
 
-        env_mod.extend(unconditional_environment_modifications(
+        env_mod.extend(uenv.unconditional_environment_modifications(
             self.default_view).reversed())
 
         for _, spec in self.concretized_specs():
             if spec in self.default_view and spec.package.installed:
                 env_mod.extend(
-                    environment_modifications_for_spec(
+                    uenv.environment_modifications_for_spec(
                         spec, self.default_view).reversed())
         return env_mod.shell_modifications(shell)
 
@@ -1461,68 +1461,6 @@ class Environment(object):
         deactivate()
         if self._previous_active:
             activate(self._previous_active)
-
-
-prefix_inspections = {
-    'bin': ['PATH'],
-    'lib': ['LD_LIBRARY_PATH', 'LIBRARY_PATH'],
-    'lib64': ['LD_LIBRARY_PATH', 'LIBRARY_PATH'],
-    'man': ['MANPATH'],
-    'share/man': ['MANPATH'],
-    'share/aclocal': ['ACLOCAL_PATH'],
-    'include': ['CPATH'],
-    'lib/pkgconfig': ['PKG_CONFIG_PATH'],
-    'lib64/pkgconfig': ['PKG_CONFIG_PATH'],
-    '': ['CMAKE_PREFIX_PATH']
-}
-
-if sys.platform == 'darwin':
-    for subdir in ('lib', 'lib64'):
-        prefix_inspections[subdir].append('DYLD_LIBRARY_PATH')
-
-
-def unconditional_environment_modifications(view):
-    """List of environment (shell) modifications to be processed for view.
-
-    This list does not depend on the specs in this environment"""
-    env = spack.util.environment.EnvironmentModifications()
-
-    for subdir, vars in prefix_inspections.items():
-        full_subdir = os.path.join(view.root, subdir)
-        for var in vars:
-            env.prepend_path(var, full_subdir)
-
-    return env
-
-def environment_modifications_for_spec(spec, view=None):
-    """List of environment (shell) modifications to be processed for spec.
-
-    This list is specific to the location of the spec or its projection in
-    the view."""
-    spec = spec.copy()
-    if view:
-        spec.prefix = Prefix(view.view().get_projection_for_spec(spec))
-
-    # generic environment modifications determined by inspecting the spec
-    # prefix
-    env = spack.util.environment.inspect_path(
-        spec.prefix,
-        prefix_inspections,
-        exclude=spack.util.environment.is_system_path
-    )
-
-    # Let the extendee/dependency modify their extensions/dependents
-    # before asking for package-specific modifications
-    env.extend(
-        build_env.modifications_from_dependencies(
-            spec, context='run'
-        )
-    )
-    # Package specific modifications
-    build_env.set_module_variables_for_package(spec.package)
-    spec.package.setup_run_environment(env)
-
-    return env
 
 
 def display_specs(concretized_specs):
